requirements.txt
ckanext-toolbelt~=0.3.10
citeproc-py==0.6.0
